Reorganized a little bit
