+++ crossbuild/crossbuild.go
@@ -12,16 +12,15 @@ var (
 	// ErrMsgEnvironmentInitFailed occurs if environment initialization failed.
 	ErrMsgEnvironmentInitFailed = "failed to initialize crossbuild environment"
 
-	// Path to crossbuild home directory inside OPS home directory.
-	crossBuildHomeDirPath = filepath.Join(lepton.GetOpsHome(), "crossbuild")
+	// CrossBuildHomeDirPath is path to crossbuild home directory.
+	CrossBuildHomeDirPath = filepath.Join(lepton.GetOpsHome(), "crossbuild")
 )
 
 func init() {
 	directories := []string{
-		crossBuildHomeDirPath,
-		vmImageDirPath,
+		CrossBuildHomeDirPath,
+		VMImageDirPath,
 	}
-
 	for _, dir := range directories {
 		if _, err := os.Stat(dir); os.IsNotExist(err) {
 			if err = os.MkdirAll(dir, 0755); err != nil {
@@ -30,7 +29,7 @@ func init() {
 		}
 	}
 
-	if _, err := os.Stat(configFilePath); os.IsNotExist(err) {
+	if _, err := os.Stat(ConfigFilePath); os.IsNotExist(err) {
 		conf := &Configuration{}
 		if err := conf.Save(); err != nil {
 			log.Fail(ErrMsgEnvironmentInitFailed, err)
