+++ crossbuild/env.go
@@ -2,6 +2,7 @@ package crossbuild
 
 import (
 	"encoding/json"
+	"errors"
 	"io/fs"
 	"io/ioutil"
 	"os"
@@ -21,11 +22,10 @@ const (
 
 // Environment is a crossbuild configuration.
 type Environment struct {
-	ID   string `json:"id"`
-	Name string `json:"name"`
-
-	vm     *virtualMachine `json:"-"`
-	source source          `json:"-"`
+	ID     string `json:"id"`
+	Name   string `json:"name"`
+	vm     *VM    `json:"-"`
+	source source `json:"-"`
 }
 
 // Remove removes environment from local directory and VM.
@@ -185,8 +185,63 @@ func LoadEnvironment() (*Environment, error) {
 	if err != nil {
 		return nil, err
 	}
-	vm.workingDirPath = env.vmDirPath()
+	vm.WorkingDirPath = env.vmDirPath()
 	env.vm = vm
 
 	return env, nil
+}
+
+// InstallDependencies installs configured source dependencies in VM.
+func (env *Environment) InstallDependencies() error {
+	if len(env.source.Dependencies) == 0 {
+		return errors.New("there are no dependencies configured")
+	}
+
+	for _, dep := range env.source.Dependencies {
+		if dep.Type == "package" {
+			if dep.Name == "nodejs" || dep.Name == "npm" {
+				prepCmd := env.vm.NewCommand(
+					"apt-get install curl software-properties-common -y",
+				).Then("apt-get update -y").AsAdmin()
+				if err := prepCmd.Execute(); err != nil {
+					return err
+				}
+
+				dep.Name = "nodejs"
+			}
+
+			vmCmd := env.vm.NewCommand("apt-get", "install", dep.Name, "-y").AsAdmin()
+			if err := vmCmd.Execute(); err != nil {
+				return err
+			}
+			continue
+		}
+
+		if dep.Type == "command" {
+			vmCmd := env.vm.NewCommandf("cd %s && %s", env.vmDirPath(), dep.Command)
+			if dep.AsAdmin {
+				vmCmd.AsAdmin()
+			}
+
+			if err := vmCmd.Execute(); err != nil {
+				return err
+			}
+			continue
+		}
+	}
+	return nil
+}
+
+// Run executes run command inside VM.
+func (env *Environment) Run() error {
+	if env.source.Commands.Run == "" {
+		return errors.New("run command is not configured")
+	}
+
+	if err := env.Sync(); err != nil {
+		return err
+	}
+
+	vmCmd := env.vm.NewCommandf("cd %s && %s", env.vmDirPath(), env.source.Commands.Run)
+	return vmCmd.Execute()
 }
